WIP: remove clovers-svc-common, repeat yourself in the two crates instead. cleanup more later
