+++ Cargo.toml
@@ -4,5 +4,4 @@ resolver = "2"
 members = [
   "clovers-back",
   "clovers-batch",
-  "clovers-svc-common",
 ]
