+++ clovers-back/Cargo.toml
@@ -8,12 +8,12 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.57"
 axum = "0.5.6"
-clovers-svc-common = { path = "../clovers-svc-common" }
+clovers = { git = "https://github.com/Walther/clovers/", rev = "b49c9c2203b79667ff70ebf6af840bdb492a0875", default-features = false, features = ["serde-derive", "stl"] }
 dotenv = "0.15.0"
 redis = { version = "0.21.5", features = ["tokio-comp", "connection-manager"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.68"
-sqlx = { version = "0.5", features = [ "postgres", "runtime-tokio-rustls", "json" ] }
+sqlx = { version = "0.5", features = [ "postgres", "runtime-tokio-rustls", "uuid", "json" ] }
 tokio = { version = "1.0", features = ["full"] }
 tower-http = { version = "0.3.3", features = ["trace", "cors"] }
 tracing = "0.1"
