+++ clovers-back/src/store.rs
@@ -1,5 +1,3 @@
-use std::str::FromStr;
-
 use super::RenderTask;
 use anyhow::{anyhow, Result};
 use redis::{aio::ConnectionManager, AsyncCommands};
@@ -86,57 +84,6 @@ WHERE id = ( $1 )
     Ok(render_task.as_ref().to_owned())
 }
 
-/// Deletes the full rendering task by id.
-pub async fn delete_render_task(id: Uuid, postgres_pool: &Pool<Postgres>) -> Result<Uuid> {
-    let id: Uuid = match sqlx::query(
-        r#"
-DELETE FROM render_tasks
-WHERE id = ( $1 )
-RETURNING id
-        "#,
-    )
-    .bind(id)
-    .fetch_one(postgres_pool)
-    .await
-    {
-        Ok(row) => row.try_get("id")?,
-        Err(e) => {
-            let error_message = format!("Error saving rendertask to postgres: {e}");
-            tracing::error!("{error_message}");
-            return Err(anyhow!("{error_message}"));
-        }
-    };
-
-    Ok(id)
-}
-
-/// Saves the full rendering result
-pub async fn save_render_result(
-    render_result: RenderResult,
-    postgres_pool: &Pool<Postgres>,
-) -> Result<Uuid> {
-    let id: Uuid = match sqlx::query(
-        r#"
-INSERT INTO render_results ( data )
-VALUES ( $1 )
-RETURNING id
-        "#,
-    )
-    .bind(render_result.data)
-    .fetch_one(postgres_pool)
-    .await
-    {
-        Ok(row) => row.try_get("id")?,
-        Err(e) => {
-            let error_message = format!("Error saving rendertask to postgres: {e}");
-            tracing::error!("{error_message}");
-            return Err(anyhow!("{error_message}"));
-        }
-    };
-
-    Ok(id)
-}
-
 /// Gets the full rendering result by id.
 pub async fn get_render_result(id: Uuid, postgres_pool: &Pool<Postgres>) -> Result<Vec<u8>> {
     let data: Vec<u8> = match sqlx::query(
@@ -188,11 +135,4 @@ SELECT id FROM render_results
     };
 
     Ok(render_result_ids)
-}
-
-/// Pops the first rendering task in the rendering queue, returning the id.
-pub async fn pop_render_queue(redis_connection: &mut ConnectionManager) -> Result<Uuid> {
-    let rendertask_id: String = redis_connection.lpop(RENDER_QUEUE_NAME, None).await?;
-    let rendertask_id = Uuid::from_str(&rendertask_id)?;
-    Ok(rendertask_id)
 }
