+++ clovers-batch/stl/teapot.stl
