--- clovers-svc-common/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "clovers-svc-common"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.57"
-dotenv = "0.15.0"
-redis = { version = "0.21.5", features = ["tokio-comp", "connection-manager"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.68"
-sqlx = { version = "0.5", features = [ "postgres", "uuid", "runtime-tokio-rustls", "json" ] }
-tracing = "0.1"
-clovers = { git = "https://github.com/Walther/clovers/", rev = "b49c9c2203b79667ff70ebf6af840bdb492a0875", default-features = false, features = ["serde-derive", "stl"] }
