+++ docker-compose.yml
@@ -3,8 +3,7 @@ services:
   clovers-back:
     restart: always
     build:
-      context: .
-      dockerfile: ./clovers-back/Dockerfile
+      context: ./clovers-back/
     ports:
       - 8080:8080
     environment:
@@ -20,8 +19,7 @@ services:
   clovers-batch:
     restart: always
     build:
-      context: .
-      dockerfile: ./clovers-batch/Dockerfile
+      context: ./clovers-batch/
     environment:
       - POSTGRES_CONNETIONINFO=postgres://postgres:local_development_passphrase@postgres/
       - REDIS_CONNETIONINFO=redis://redis:6379/
